starter-labs/llm-sec-lab4-starter/llm-sec-lab4/requirements.txt
python-dotenv>=1.0.1
google-genai>=1.0.0
pydantic>=2.9.0
pyyaml>=6.0.2
